Fix bugs from previous merges
