+++ app/build.gradle
@@ -32,6 +32,8 @@ dependencies {
     androidTestImplementation 'androidx.test:runner:1.1.0-alpha4'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.1.0-alpha4'
     implementation 'com.squareup.retrofit2:retrofit:2.5.0'
+
+    implementation 'com.google.android.material:material:1.1.0-alpha05'
     implementation 'com.squareup.retrofit2:converter-gson:2.4.0'
     implementation 'com.squareup.okhttp3:logging-interceptor:3.12.1'
     implementation 'com.jakewharton:butterknife:10.1.0'
