+++ app/src/main/AndroidManifest.xml
@@ -15,19 +15,18 @@
         <activity android:name=".activities.ForumActivity"></activity>
         <activity android:name=".RegisterActivity" />
         <activity android:name=".activities.ProfileActivity" />
-        <activity android:name=".activities.MainActivity">
+        <activity android:name=".activities.splashScreen" />
+
+        <activity android:name=".activities.LoginActivity" />
+        <activity android:name=".activities.MainActivity"
+            android:label="@string/title_activity_nav_drawer"
+            android:theme="@style/AppTheme.NoActionBar" >
+
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
-        </activity>
-        <activity android:name=".activities.splashScreen" />
-
-        <activity android:name=".activities.LoginActivity" />
-        <activity android:name=".activities.NavDrawerActivity"
-            android:label="@string/title_activity_nav_drawer"
-            android:theme="@style/AppTheme.NoActionBar" >
         </activity>
