--- app/src/main/res/layout/content_nav_drawer.xml
@@ -1,11 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<android.support.constraint.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    app:layout_behavior="@string/appbar_scrolling_view_behavior"
-    tools:context="hng.tech.apoe_4.activities.NavDrawerActivity"
-    tools:showIn="@layout/app_bar_nav_drawer">
-
-</android.support.constraint.ConstraintLayout>
